move imported tokens to a separate file
